Wake up on POLLIN and POLLOUT on the ssh socket.
This fixes a weird issue where new X11 applications will not be
able to start rendering on screen until an input even is forced
on the other X11 connections, e.g. by moving your mouse over them.

This has something to do with complications of the non-blocking I/O
in libssh2, but I'm at a loss to better describe why this make it
work correctly.